moved homepage to separate module
