+++ app/modules/home/components/HomeSearchSection.js
@@ -1,0 +1,16 @@
+// @flow
+
+import * as React from 'react';
+import { Container } from 'semantic-ui-react';
+
+const HomeSearchSection = (): React.Node => {
+  return (
+    <Container>
+      <p>
+        Search for topics here!
+      </p>
+    </Container>
+  );
+};
+
+export default HomeSearchSection;
