+++ app/modules/home/components/index.js
@@ -1,0 +1,12 @@
+// @flow
+
+import HomeContainer from './HomeContainer';
+import HomeSearchSection from './HomeSearchSection';
+import HomeSocialFeed from './HomeSocialFeed';
+
+export const components = {
+  homecontainer: HomeContainer,
+  homesearchsection: HomeSearchSection,
+  homesocialfeed: HomeSocialFeed,
+};
+
