+++ app/modules/home/index.js
@@ -1,0 +1,11 @@
+// @flow
+
+import * as components from './components';
+
+
+const home = {
+  components,
+};
+
+
+export default home;
